+++ node/training2.js
